Changing the size of the frequency variable

+++ COTS/MCAL/05-TIMERS/TIMER1/TIMER_1_program.c
@@ -122,7 +122,7 @@ Std_ReturnType Timer1_write(const Timer1_t *_timer, u16 _value)
  *          (E_OK) : The function done successfully
  *          (E_NOT_OK) : The function has issue while performing this action
  */
-Std_ReturnType Timer1_read(const Timer1_t *_timer, u16 *_value, u16 *_freq, u8 *_duty_cycle)
+Std_ReturnType Timer1_read(const Timer1_t *_timer, u16 *_value, u32 *_freq, u8 *_duty_cycle)
 {
 	Std_ReturnType ret = E_OK;
 	u16 firstCapture = 0;
